+++ src/main/java/app/ebr/repositories/ParkingLotRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import app.ebr.domains.models.ParkingLot;
 
 @Repository
-public interface ParkingLotRepository extends CrudRepository<ParkingLot, Long> {
+public interface ParkingLotRepository extends CrudRepository<ParkingLot, Integer> {
 
 }
